Add files via upload
The project titled as “Banking Automation” is a windows base application. This software is developed with scalability in mind. Additional modules can be easily added when necessary. The software is developed with modular approach to deal with managing accounts and their related operations. Tasks involved in this project are opening the user accounts, recording the account holder’s transactions, depositing & withdrawing amount. All modules in the system have been tested with valid data and invalid data and everything work successfully.  The system has fulfilled all the objectives identified and is able to replace the existing system.